requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
diffusers==0.2.4
filelock==3.8.0
ftfy==6.1.1
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dotenv==0.20.0
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
scipy==1.9.0
tokenizers==0.12.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.64.0
transformers==4.21.2
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
zipp==3.8.1
